+++ src/palletops/crate/hadoop/base.clj
@@ -98,7 +98,9 @@
      :else (let [[url md5-url] (url settings)]
              (assoc settings
                :install-strategy ::remote-directory
-               :remote-directory {:url url :md5-url md5-url})))))
+               :remote-directory {:url url
+                                  ;; pallet doesn-t like :md5-url to be nil
+                                  :md5-url (or  md5-url "")})))))
 
 (defn env-var-merge
   [a b]